refactor(package): 패키지 구조 변경
- 로또 번호 및 티켓과 관련있는 파일들을 lotto 패키지에 따로 분류
