+++ src/main/kotlin/domain/Judgement.kt
@@ -1,5 +1,8 @@
 package domain
 
+import domain.lotto.Lotto
+import domain.lotto.LottoBundle
+
 object Judgement {
     fun compareLottoBundle(winningNumbers: WinningNumbers, lottoBundle: LottoBundle): WinningResult {
         return WinningResult(lottoBundle.lottos.map { compareLotto(winningNumbers, it) })
