--- src/main/kotlin/domain/LottoBundle.kt
@@ -1,3 +1,0 @@
-package domain
-
-class LottoBundle(val lottos: List<Lotto>)
