--- src/main/kotlin/domain/LottoGenerator.kt
@@ -1,5 +1,0 @@
-package domain
-
-fun interface LottoGenerator {
-    fun generate(): Lotto
-}
