+++ src/main/kotlin/domain/lotto/Lotto.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.lotto
 
 import util.PREFIX
