+++ src/main/kotlin/domain/lotto/LottoBundle.kt
@@ -1,0 +1,3 @@
+package domain.lotto
+
+class LottoBundle(val lottos: List<Lotto>)
