+++ src/main/kotlin/domain/lotto/LottoNumber.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.lotto
 
 import util.PREFIX
