+++ src/main/kotlin/domain/lotto/generator/LottoGenerator.kt
@@ -1,0 +1,7 @@
+package domain.lotto.generator
+
+import domain.lotto.Lotto
+
+fun interface LottoGenerator {
+    fun generate(): Lotto
+}
